+++ sockeye/__init__.py
@@ -11,4 +11,5 @@
 # express or implied. See the License for the specific language governing
 # permissions and limitations under the License.
 
-__version__ = '1.18.85'
+
+__version__ = '1.18.87'