+++ sockeye/arguments.py
@@ -876,6 +876,12 @@ def add_training_args(params):
                               help='Maximum number of checkpoints the model is allowed to not improve in '
                                    '<optimized-metric> on validation data before training is stopped. '
                                    'Default: %(default)s.')
+    train_params.add_argument('--max-checkpoints',
+                              type=int,
+                              default=None,
+                              help='Maximum number of checkpoints to continue training the model '
+                                   'before training is stopped. '
+                                   'Default: %(default)s.')
     train_params.add_argument('--min-num-epochs',
                               type=int,
                               default=None,
@@ -1061,10 +1067,16 @@ def add_training_args(params):
                               choices=[C.RNN_INIT_ORTHOGONAL, C.RNN_INIT_ORTHOGONAL_STACKED, C.RNN_INIT_DEFAULT],
                               help="Initialization method for RNN parameters. Default: %(default)s.")
 
+    train_params.add_argument('--fixed-param-strategy',
+                               default=None,
+                               choices=C.FIXED_PARAM_STRATEGY_CHOICES,
+                               help="Fix various parameters during training using a named strategy. The strategy "
+                                    "name indicates which parameters will be fixed (Wuebker et al., 2018). "
+                                    "Default: %(default)s.")
     train_params.add_argument('--fixed-param-names',
                               default=[],
                               nargs='*',
-                              help="Names of parameters to fix at training time. Default: %(default)s.")
+                              help="Manually specify names of parameters to fix during training. Default: %(default)s.")
 
     train_params.add_argument(C.TRAIN_ARGS_MONITOR_BLEU,
                               default=500,