+++ sockeye/train.py
@@ -723,7 +723,8 @@ def create_training_model(config: model.ModelConfig,
                                             default_bucket_key=train_iter.default_bucket_key,
                                             bucketing=not args.no_bucketing,
                                             gradient_compression_params=gradient_compression_params(args),
-                                            fixed_param_names=args.fixed_param_names)
+                                            fixed_param_names=args.fixed_param_names,
+                                            fixed_param_strategy=args.fixed_param_strategy)
 
     return training_model
 
@@ -897,6 +898,7 @@ def train(args: argparse.Namespace) -> training.TrainState:
         if min_epochs is not None and max_epochs is not None:
             check_condition(min_epochs <= max_epochs,
                             "Minimum number of epochs must be smaller than maximum number of epochs")
+
         # Fixed training schedule always runs for a set number of updates
         if args.learning_rate_schedule:
             min_updates = None
@@ -921,6 +923,7 @@ def train(args: argparse.Namespace) -> training.TrainState:
                                      metrics=args.metrics,
                                      checkpoint_interval=args.checkpoint_interval,
                                      max_num_not_improved=max_num_checkpoint_not_improved,
+                                     max_checkpoints=args.max_checkpoints,
                                      min_samples=min_samples,
                                      max_samples=max_samples,
                                      min_updates=min_updates,